+++ NEWS.md
@@ -2,8 +2,8 @@
 
 # vcfR 1.0.0.9000
 
+* `read.vcfR()` now handles tilde expansion.
 * `addID()` populates the non-missing values in the ID column of VCF data by concatenating the chromosome and position. 
-
 
 # vcfR 1.0.0
 Released on CRAN 2016-02-22.