+++ NEWS.md
@@ -2,7 +2,8 @@
 
 # vcfR 1.0.0.9000
 
-* `maf()` provides counts and frequency for teh minor (or other) allele.
+* `write.vcf()` now uses mask = TRUE.
+* `maf()` provides counts and frequency for the minor (or other) allele.
 * `create.chromR()` now handles instances with no seq and the annotation position exceeds the greatest VCF POS.
 * `read.vcfR()` now handles tilde expansion.
 * `addID()` populates the non-missing values in the ID column of VCF data by concatenating the chromosome and position.